Disperse geo.js methods into respective files
