+++ src/store/location.js
@@ -3,7 +3,6 @@
 
 import { distance } from '@turf/distance';
 import { point } from '@turf/helpers';
-import { geoToXY } from '../utils/geo.js';
 import { computed, reactive } from 'vue';
 
 export const myLocation = reactive({
@@ -45,3 +44,37 @@ export const distanceTo = computed(() => {
     return distance(myTurfPoint.value, someLocation, options);
   };
 });
+
+function geoToXY(myLocation, myHeading, poiLocation) {
+  // Convert degrees to radians
+  const toRadians = degree => degree * (Math.PI / 180);
+
+  // Earth radius in meters
+  const earthRadius = 6371000;
+
+  // Calculate the relative distance in meters
+  const deltaLat = toRadians(poiLocation.geometry.coordinates[1] - myLocation.geometry.coordinates[1]);
+  const deltaLon = toRadians(poiLocation.geometry.coordinates[0] - myLocation.geometry.coordinates[0]);
+
+  const a =
+    Math.sin(deltaLat / 2) * Math.sin(deltaLat / 2) +
+    Math.cos(toRadians(myLocation.geometry.coordinates[1])) * Math.cos(toRadians(poiLocation.geometry.coordinates[1])) *
+    Math.sin(deltaLon / 2) * Math.sin(deltaLon / 2);
+
+  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+  const distance = earthRadius * c;
+
+  // Calculate the angle between the Y-axis and the line connecting myLocation and poiLocation
+  const angleToTarget = Math.atan2(
+    poiLocation.geometry.coordinates[0] - myLocation.geometry.coordinates[0],
+    poiLocation.geometry.coordinates[1] - myLocation.geometry.coordinates[1]
+  ) - toRadians(myHeading);
+
+  // Calculate X and Y coordinates
+  const x = distance * Math.sin(angleToTarget);
+  const y = distance * Math.cos(angleToTarget);
+
+  // Scale so that sounds are more audible, and to fit on canvas
+  const scaleFactor = 0.05;
+  return { x: x * scaleFactor, y: y * scaleFactor };
+}
